requirements.txt
Flask==2.3.2
google-generativeai
colorama==0.4.6